Added Intake Sweeper method
Added an intake sweeper method as well as everything that goes with it.

Still haven't figured out why my android studio doesn't have the proper licensing so I haven't run this through a compiler yet so unknown if there are any errors. Likely there are some but problems shouldn't be big and easily fixable.